Moves the profiling to its own new manager, profile manager. Adds a button to the debug console, to run a local profile of 30 frames.
