+++ UnityProject/Assets/IngameDebugConsole/Scripts/DebugLogManager.cs
@@ -589,6 +589,22 @@ namespace IngameDebugConsole
 		}
 
 		/// <summary>
+		/// Creates a local profile
+		/// </summary>
+		public void LocalProfileButtonPressed()
+		{
+			if (Debug.isDebugBuild)
+			{
+				ProfileManager.Instance.StartProfile(30);
+				Logger.Log("Running a local profile, saving on installation folder");
+			}
+			else
+			{
+				Logger.Log("Unable to run local profile, the build needs to be in development mode");
+			}
+		}
+
+		/// <summary>
 		/// Debug window is being resized,
 		/// Set the sizeDelta property of the window accordingly while
 		/// preventing window dimensions from going below the minimum dimensions
