+++ UnityProject/Assets/Resources/Prefabs/SceneConstruction/NestedManagers/ProfileManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 144ea30de3c7cfa498603ec000185d6c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
