+++ UnityProject/Assets/Scripts/Managers/ProfileManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e43f2e76098a24047910dd42cfe91338
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
